Move nextpow2 to tftb.utils
It's a function used by both generators and processors, no sense in
having it inside the generators module.
