+++ tftb/generators/api.py
@@ -1,7 +1,7 @@
 from amplitude_modulated import (amexpos, amgauss, amrect, amtriang)
 from frequency_modulated import (fmconst, fmhyp, fmlin, fmodany, fmpar,
         fmpower, fmsin)
-from utils import nextpow2, sigmerge
+from utils import sigmerge
 from noise import dopnoise, noisecg, noisecu
 from analytic_signals import (anaask, anabpsk, anafsk, anapulse, anaqpsk,
         anasing, anastep, hilbert)
